// mobile/src/types/domain.ts
export interface User {
  id: string;
  email: string;
  name: string;
  avatarUrl?: string | null;
  city?: string | null;
  createdAt: string;
  updatedAt: string;
}